requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
openai
pillow
requests
aiohttp
python-multipart
typing
chromadb
numpy